+++ lib/_internal/_p2a/index.d.ts
