+++ lib/_internal/_p2a/index.js
