+++ lib/_internal/_placehold/index.js
@@ -1,5 +1,5 @@
-import __ from '../__'
-import toArray from '../toArray'
+import __ from '../../__'
+import toArray from '../../toArray'
 
 export default function placehold(fn, length, oldArgs) {
   oldArgs = oldArgs || []
