+++ lib/_internal/_placehold/placehold.test.js
@@ -1,4 +1,4 @@
-var __ = require('../__')
+var __ = require('../../__')
 var placehold = require('.')
 
 var first = (a, b, c) => a
