+++ lib/curryN/index.js
@@ -1,4 +1,4 @@
-import placehold from '../placehold'
+import placehold from '../_internal/placehold'
 
 function curryN(n, fn) {
   return placehold(function() {
