+++ lib/flattenObj/index.js
@@ -1,5 +1,5 @@
-import a2p from '../a2p'
-import p2a from '../p2a'
+import a2p from '../_internal/a2p'
+import p2a from '../_internal/p2a'
 import nAry from '../nAry'
 
 export default nAry(1, function objToFlat(obj, nest) {
