+++ lib/lensPath/index.js
@@ -1,5 +1,5 @@
 import lens from '../lens'
-import p2a from '../p2a'
+import p2a from '../_internal/p2a'
 
 export default function lensPath(k) {
   var p = p2a(k)
